up more

+++ src/diffusers/models/transformers/transformer_ltx.py
@@ -508,6 +508,5 @@ def apply_rotary_emb(x, freqs):
     cos, sin = freqs
     x_real, x_imag = x.unflatten(2, (-1, 2)).unbind(-1)  # [B, S, H, D // 2]
     x_rotated = torch.stack([-x_imag, x_real], dim=-1).flatten(2)
-    # YiYi TODO: testing only, remove this change before merging
-    out = (x * cos.to(x.dtype) + x_rotated * sin.to(x.dtype)).to(x.dtype)
+    out = (x.float() * cos + x_rotated.float() * sin).to(x.dtype)
     return out
